requirements.txt
discord.py==2.3.2
Pillow==10.0.1
requests==2.31.0
aiohttp==3.8.5
aiodns==3.0.0